feat: 내부 Enum -> 공통 Enum 변경
