+++ payment-service/src/main/java/com/javaauction/payment_service/domain/enums/HoldStatus.java
@@ -1,0 +1,5 @@
+package com.javaauction.payment_service.domain.enums;
+
+public enum HoldStatus {
+    HOLD_ACTIVE, HOLD_RELEASED, HOLD_CAPTURED
+}
