+++ payment-service/src/main/java/com/javaauction/payment_service/domain/enums/TransactionType.java
@@ -1,0 +1,5 @@
+package com.javaauction.payment_service.domain.enums;
+
+public enum TransactionType {
+    CHARGE, WITHDRAW, PAYMENT, HOLD
+}
