+++ payment-service/src/main/java/com/javaauction/payment_service/domain/model/WalletTransaction.java
@@ -1,5 +1,8 @@
 package com.javaauction.payment_service.domain.model;
 
+import com.javaauction.payment_service.domain.enums.ExternalType;
+import com.javaauction.payment_service.domain.enums.HoldStatus;
+import com.javaauction.payment_service.domain.enums.TransactionType;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
@@ -18,16 +21,4 @@ public class WalletTransaction {
     private final HoldStatus holdStatus;
     private final ExternalType externalType;
     private final UUID externalId;
-
-    public enum TransactionType {
-        CHARGE, WITHDRAW, PAYMENT, HOLD
-    }
-
-    public enum HoldStatus {
-        HOLD_ACTIVE, HOLD_RELEASED, HOLD_CAPTURED
-    }
-
-    public enum ExternalType {
-        AUCTION, BID
-    }
 }
