+++ payment-service/src/main/java/com/javaauction/payment_service/infrastructure/persistence/entity/WalletTransactionEntity.java
@@ -1,6 +1,9 @@
 package com.javaauction.payment_service.infrastructure.persistence.entity;
 
 import com.javaauction.global.infrastructure.entity.BaseEntity;
+import com.javaauction.payment_service.domain.enums.ExternalType;
+import com.javaauction.payment_service.domain.enums.HoldStatus;
+import com.javaauction.payment_service.domain.enums.TransactionType;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -38,17 +41,4 @@ public class WalletTransactionEntity extends BaseEntity {
     private ExternalType externalType;
 
     private UUID externalId;
-
-    public enum TransactionType {
-        CHARGE, WITHDRAW, PAYMENT, HOLD
-    }
-
-    public enum HoldStatus {
-        HOLD_ACTIVE, HOLD_RELEASED, HOLD_CAPTURED
-    }
-
-    public enum ExternalType {
-        AUCTION, BID
-    }
-
 }
