+++ payment-service/src/main/java/com/javaauction/payment_service/presentation/dto/response/WalletTransactionDto.java
@@ -1,5 +1,6 @@
 package com.javaauction.payment_service.presentation.dto.response;
 
+import com.javaauction.payment_service.domain.enums.TransactionType;
 import com.javaauction.payment_service.domain.model.WalletTransaction;
 import lombok.Builder;
 import lombok.Getter;
@@ -11,7 +12,7 @@ import java.util.UUID;
 public class WalletTransactionDto {
     private UUID walletTransactionId;
     private Long amount;
-    private WalletTransaction.TransactionType transactionType;
+    private TransactionType transactionType;
 
     public static WalletTransactionDto from(WalletTransaction walletTransaction) {
         return WalletTransactionDto.builder()
